Split up the `CheckBuildNum` class for each individual server platform to avoid runtime import resolution failures due to needlessly attempting to import classes that ultimately go unused and may not even _exist_ on some server platforms. This change also makes makes GeyserUpdate fully function on every Minecraft version between 1.8 and 1.16.5!
