+++ src/main/java/com/alysaa/geyserupdater/spigot/command/GeyserCommand.java
@@ -1,7 +1,8 @@
 package com.alysaa.geyserupdater.spigot.command;
 
-import com.alysaa.geyserupdater.common.util.CheckBuildNum;
 import com.alysaa.geyserupdater.spigot.SpigotUpdater;
+import com.alysaa.geyserupdater.spigot.util.GeyserSpigotCheckBuildNum;
+
 import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
@@ -15,11 +16,11 @@ public class GeyserCommand implements CommandExecutor {
             Player player = (Player) sender;
             if (command.getName().equalsIgnoreCase("geyserupdate") && player.hasPermission("gupdater.geyserupdate")) {
                 sender.sendMessage(ChatColor.WHITE + "[GeyserUpdater] Checking current Geyser version!");
-                CheckBuildNum.checkBuildNumberSpigot();
+                GeyserSpigotCheckBuildNum.checkBuildNumberSpigot();
             }
         } else if (sender instanceof ConsoleCommandSender) {
             SpigotUpdater.plugin.getLogger().info("Checking current Geyser version!");
-            CheckBuildNum.checkBuildNumberSpigot();
+            GeyserSpigotCheckBuildNum.checkBuildNumberSpigot();
         }
         return false;
     }
